debug: remove "debug find" (#292)
This command was broken since #260, because it relies on matching by
name instead of ID.

Since #260 we match simple shards by name and repos within compound
shards by ID. For this command to work we would have to give the ID and
the name as an argument, which is inconvenient.

I have been using a combination of "debug meta" with "jq" since #260 and
it works fine for me. I think we should remove this command for now. We
might bring it back once we match everything by ID.