ci: run tests for PRs from forks (#306)
So far we only ran tests for pushes. However, for external contributers opening PRs from forks, GitHub doesn't send "push" events but "pull_request" events (and variants thereof).

With this change we run tests for all PRs, including subsequent pushes to the the same PR. Once the PR is merged to master we run test and build just like before.