ranking: enhance scoring based on ctags (#296)
ctags provides information, such as position, language, and kind, about symbols. For scoring we only use the posititon to increase the score of matches that are contained within a symbol.

With this change we introduce a scoring function that boosts results based on the kind of the symbol. The core idea is that, depending on the language, some language entities are more important than others. EG based on user feedack, classes in Java are considered to be more relevant than instance variables, given both match the query.

For now the scoring function is just a skeleton targeting one language.

Co-authored-by: Keegan Carruthers-Smith <keegan.csmith@gmail.com>