zoekt-sourcegraph-indexserver: add support for preparing delta builds behind a feature flag
This PR adds an environment variable named `DELTA_BUILD_REPOS_ALLOWLIST`, which is a comma separated list of repositories that zoekt should try using the new delta build strategy for (as described in https://github.com/sourcegraph/sourcegraph/issues/29731).

For all such repositories, zoekt-sourcegraph-indeserver will now fetch both the current branch commits and the commits that were used to generate the shards from the most recent index. If any aspect of this process fails (such as if those older commits are no longer available), zoekt-sourcegraph-indexserver will still proceed. However, it'll rely on zoekt-git-index to fall back to a "normal" build.