builder.go: fallback to normal build if a sourcegraph ignore file has been modified
I also discovered that go-git has really unintuitive handling of file names that come from diffs when the files are in a folder. After a lot of trial and error, I eventually stumbled-upon a "hack" to always get the full path relative to the repository's root: https://stackoverflow.com/a/58922399

I also added an extra test to make sure that this doesn't break in the future.