feature flag lazy runeDocSection decoding (#312)
Currently this takes up 30% of production CPU. We are unsure if its
worth the tradeoff so we will experiment with enabling vs disabling and
measure the impact.

We default to the current behaviour. To disable you set the environment
variable ZOEKT_DISABLE_LAZY_DOC_SECTIONS to a non empty value.

Test Plan: locally run zoekt-webserver with and without the feature
flag. With the feature flag we should have higher memory usage. Ensure
in both cases symbol search works.