zoekt-hg-index: remove due to broken zstd dep
Our CI is failing due to the HG library we use depending on zstd. This
is the easiest fix since I am unaware of users of zoekt-hg. If you come
across this commit, please complain and I will see what I can do to fix
it.

Test Plan: CI works.