indexeserver: better logging for sourcegraphClient (#341)
This improves the error messages for our sourcegraph client. The change is motivated by a recent customer issue.

Now we log the error returned by Sourcegraph if the List call returns with status code 500. We also wrap errors coming from Server.Sourcegraph.List(...), which should make it easier to filter the logs.