fix relative links on site admin proxy view
The site admin view in Sourcegraph requires the services to handle a
request at their respective root "/". Hence we move the re-index view
back to the root.

At the same time we fix a bug in production. Clicking on re-index
when comming from the Sourcegraph site admin view sent the form to the
absolute instead of the relative root.