ranking: boost base filename matches (#349)
We boost matches of filenames if the match is within the boundaries of
the base of a filename. Full matches get a higher boost than partial
matches. The approach is identical to boosting of content matches that
match symbol section.

Additionally we fix a bug where we used sections belonging to the file
content to score filename matches.

Co-authored-by: Keegan Carruthers-Smith <keegan.csmith@gmail.com>