tests: fix alignment for number hints like "--0123"
Some tests include queries that match somewhere in the middle
of a content string. For a programmer reading or debugging these
tests, they contain visual aids that assume a fixed-width font
and look like this

  var contentString = "abcd"
  // ------------------0123

This commit restores the alignment of these helping comments that
was broken by multiple refactorings, adds the hints where they could
help, and removes the hints where they cannot be written in this way
or where the offset does not matter.