Revert "Revert "all: use go1.18 (#336)" (#373)"
This reverts commit d20ed38dd968c6ad94cba8ed72376cfb45c1ac1f.

We have established that the upgrade to go 1.18 led to a higher reported
numbers for MEM in production. Specifically
`cadvisor_container_memory_usage_percentage_total` increases
significantly.

The go 1.18 release notes mention a change in the garbage collector
which might have an impact on memory. "The intended workaround is to
tweak GOGC where necessary.".

Our strategy is to upgrade to go 1.18 an establish whether tweaking GOGC
has the intended effect.