docker: change our default GOGC to 25 (#399)
We had a regression in go1.18 in our memory usage. From experimentation
we have found a value of 25 gave us the same stable size from before
1.18.

Note: we set GOGC for all containers, but really we only should be doing
this for webserver. However, I don't want to bundle that change into
this one.

Test Plan: we have validated this on sourcegraph.com over the last few
weeks.