gitlab: Iterate over projects by id instead of offset based pagination
There is a limit for offset based pagination on the projects API. One
option is to use keyset's but go-gitlab has no support for it. Another
related approach is to iterate over project ids. For this we are going
to:
 * Order by 'id'
 * Sort in ascending order
 * Use ID After to move forward until there are no ids

This helps to iterate the number of projects on a large(r) instance
of GitLab.