webserver: do not wait until all shards loaded to be available
Right now on large instances a zoekt-webserver can take on the order of
minutes to load all shards on startup. This means a whole chunk of our
search index is completely unavailable during that process. This is
excerbated by kubernetes' statefulsets which only roll out one shard at
a time, meaning for a long time there is partial availability.

This change adjusts the webserver to declare it is ready straight away
and instead load in shards async. This means we will more quickly
rollout, overall reducing partial availability of the cluster.

Additionally we adjusted the loader to publish what is has loaded so far
every 5 seconds.

Test Plan: go test covers that we are indeed still blocking in the
normal call to NewDirectorySearcher. To test NewDirectorySearcherFast we
locally ran zoekt-webserver with extra logging. It showed that it was
ready before shards were loaded, and searching still worked.
Additionally we introduced artifical delay in shard loading and observed
partial availability in search results (but no errors).

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>