ranking: boost Java matches based on symbol kind (#421)
With this change we add more granular boosts for Java matches based on
their "kind".

I am also correcting a inconsequential mistake I made earlier. Since we
use "interactive" mode in go-ctags, the output format is JSON which sets
the "kind" to its long form and not the the single letter form. For
example "class" instead of "c". Therfore we can remove the one-letter
cases from scoring.

Thanks to @olafurpg for ordering the Java "kinds" by importance.