api: estimate size of zoekt.SearchResult
We add `SizeBytes` as method to `zoekt.SearchResult` which
estimates the size of a search result on a best effort basis.

New unit tests make sure we catch regressions.

Note, we use `int` in various places. For the estimates I assume we a running
on 64-bit architecture. Alignments are ignored.

The goal is to use this estimate to define a threshold to flush the
result queue in Sourcegraph.