api: estimate size of zoekt.SearchResult (#425)
We add `zoekt.SearchResult.SizeBytes()` which estimates the size of a search result on a best-effort basis.

The goal is to use this estimate to decide when to flush the result queue in Sourcegraph.

Note: We use `int` in various places. For the estimates I assume we a running on a 64-bit architecture. Alignments are ignored.