all: remove bloom filter code
We have had this disabled on the read path for many months. Additionally
I noticed in profiles of zoekt-git-index that generation of the bloom
filter was taking 24% of the time on the sourcegraph repo.

Test Plan: Tested the matrix of old/new zoekt-webserver reading old/new
shards to ensure everything just keeped working nicely.