zoekt-git-index: add cpuprofile and memprofile flags (#429)
This makes it possible to get profiles for indexing with git. The code
is copy-pasta from the pprof godoc. Note: we move main into a function
which returns such that the defers run before os.Exit is called.

Also made it so we don't use parallelism if we are memory profile.