shards: introduce SearchOptions.FlushWallTime
This option makes it so that streaming search will collect results for
FlushWallTime before sending. This allows a client to provide
best-effort ranking based on time. We hope this will be a sweet spot
between the benefits of streaming latency and relevance of results.

Test Plan: go test. TODO add unit tests exercising this specific logic.