merging: delete code that removed duplicates
Maxim added this code a while ago to remove duplicate indexes in
compound shards on .com. The code path was active for a short period of time
and fulfilled its purpose. Now we can remove it.

Background:
For a while we saw a handfull of duplicate indexes per shard on .com.
The root cause turned out to be a missing lock on index dir during
merging.  This has since been fixed. However, the duplicates still
lingered which is why we added this cleanup code.

I have checked several times over the last couple of weeks and can
confirm that no new duplicates materialized.

I believe the issue is resolved.