ranking: sort docs before adding them to builder
This moves the sorting logic for offline ranking from builder to
gitindex. This means we add documents to builder in decreasing order of
their rank instead of in lexicographic order and can skip sorting later.

As a consequence, the order is now consistent accross and within shards.
This is a nice property because

(1) now we stream results in decreasing order of importance
(2) if we stop searching a repo because we already found enough matches
in some of its shards, we are sure that we skip only less important
results.

Example:

let a(x) be a document with name "a" and rank x (float64) and let's
assume that all documents have the same size and only two fit in one
shard.

Corpus:

a(0.6)
b(0.2)
c(0)
d(1.0)

Distribution of documents across the 2 shards

Before (lexicographic > rank):

Shard 1:
a(0.6)
b(0.2)

Shard 2:
d(1.0)
c(0)

Now (rank):

Shard 1:
d(1.0)
a(0.6)

Shard 2:
b(0.2)
c(0)