Add URL Template for Gitea (#453)
This adds a URL template for gitea >= 1.17.0.

The "display=source" query parameter was introduced in gitea 1.17.0 and
is required to disable file rendering. Line numbers are disabled in
rendered files, so you wouldn't be able to jump to a line without
"display=source".

The "web-url-type = gitea" config option must be set in the bare
repositories indexed with "zoekt-git-index" e.g. in
owner/repo.git/config:

  [zoekt]
    web-url-type = gitea
    web-url = https://gitea.local/owner/repo
    name = owner/repo

Note: already indexed repos must be deleted before reindexing in order
to pick up the indexer changes.