streaming search that aggregates and ranks for FlushWallTime (#438)
This option makes it so that streaming search will collect results for
FlushWallTime before sending. This allows a client to provide
best-effort ranking based on time. We hope this will be a sweet spot
between the benefits of streaming latency and relevance of results.

Additionally we also make it so that streaming search respects
MaxDocDisplayCount.

Test Plan: go test

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>