indexserver: backoff repo if failing repeatedly (#442)
Introduce backoff strategy for failed indexing attempts. The strategy's goal is to backoff subsequent indexing attempts for a given repository for a period of time because we've identified that the preceding indexing attempt has failed. The backoff period of time is finite and increases linearly with consecutive indexing failures.

Test Plan: Add queue tests and backoff duration tests