ranking: boost camelCase
With this change we recognize changes in capitalization as word boundaries.

There is a corner case which was caught in the tests, where we falsy
classify "nnerClass" as word, because the preceeding letter in the code
is a capital "I". However, I don't believe this corner case is relevant
in practive and by accepting it we keep the code simple.