add generic ctags ranking
We add a few things in this commit. Namely we add generic ranking which
is based on best guesses on what is important across languages. I
believe this is a useful change since it is mostly true and is likely
better than just treating all symbols as equal.

Secondly I add language specific ranking for python, ruby and php. These
are as usualy best guesses based on experience in the language.

Lastly our test cases for TestReadSearch had the language stored in
lower case. I noticed this thanks to the generic scorer running since
previously the go one was not. Just to be safe I made the switch
statements match against both the expected casing and lowered case of
languages.

Test Plan: go test