ranking: parametrize contribution of document ranks
We scale the contribution of document ranks to the final ranking with a
damping factor. The zero value of the damping factor leads to an equal
contribution of scores and ranks, while a damping factor of 1
effectively eliminates the influence of document ranks on the final
ranking.

The default is 0 IE equal contribution.