all: proxy requests to /indexserver/ to socket
With this change webserver proxies request to /indexserver/ to the unix
domain socket at `<index dir>/indexserver.socket`. Since this is a
Sourcegraph feature, we put the proxy behind a flag which is toggled off
per default.

The motivation is to expose indexserver endpoints on webserver
addressses, which in case of Sourcegraph, are easily accessible at
runtime.