webserver: proxy requests to indexserver via socket (#487)
With this change, webserver proxies request to `<webserver>/indexserver/` to the unix domain socket at `<index dir>/indexserver.socket`. Since this is a Sourcegraph feature, we put the proxy handler in webserver behind a flag which is toggled off per default.

The motivation is to expose indexserver endpoints on webserver addresses, which in case of Sourcegraph, are easily accessible at runtime.