shards: introduce loaded struct (#491)
I want to introduce some extra state were the order we access it is
important, so we introduce a new structure loaded which represents the
final computed state. For now thiis is a non-functional change, we only
introduce the structure and update call sites. The next commit will add
a field.

Test Plan: go test