webserver: use B+-tree to map ngrams to posting lists (#522)
Our in-memory map of ngrams to posting lists contributes about 33% to webserver's heap usage in a production setting. This replaces the current implementation with a B+-tree. We expect to reduce the memory footprint of the mapping by around 95% at the cost of slower search speeds. The change is behind a feature-flag. To activate, set  ZOEKT_ENABLE_BTREE=true in ENV.