indexserver: expose all merge options
With this change we expose "the maximum allowed priority" and "the minimum age
of the latest commit" as command line flags and ENVs.

At the same time we clean up the Server struct a bit and bundle all
fields related to shard merging in a dedicated options struct.

The motiviation is to make shard merging fully configurable before
rolling it out to all customers.