zoekt: read posting offset from disk
The btree still referenced the offsets to the posting lists which had a
big impact on the heap usage of webserver.

With this change we just hold on to the simple section pointing to the
offsets IE we read offsets lazily and therfore incur another disk
access.

I added a unit test for ngramIndex.Get, because there wasn't really a
good test so far that captured the expected size of a posting list.

In total, after this change, the retrieval of a posting list for a given
ngram via the btree requires 2 disk accesses (1 for getting the bucket
ngrams, 1 for getting the offset to the posting list).

If this turns out to be to costly with regards to latency, we should
consider storing the offset together with the ngrams in the bucket.