Rename dataDir argument in zoekt-dynamic-indexserver to repoDir
This is to avoid confusion where DATA_DIR is being used in
`Dockerfile.webserver` as the index dir. This will cause confusion
because we are adding a new `Dockerfile.dynamic-indexserver`. In that
case we'll want to use consistent env var names and it will be confusing
that `$DATA_DIR` is being passed as `index_dir` and not `data_dir`. So
to avoid that confusion we remove the concept of `data_dir` argument
altogether and rename it to `repo_dir` since it is only used for cloning
repos anyway.