cmd/zoekt-sourcegraph-indexserver: Add default port to root URL
Add a default port (80 for HTTP, 443 for HTTPS) to the root URL if one is
not explicitly specified. This ensures that clients will be able to
successfully connect when a port is omitted in configuration.

This is necessary because gRPC does not assume a default port like Go's
http client does. By always specifying a port in the root URL, the same
URL will now work with both gRPC and http clients.

A new addDefaultPort helper function is added and used to handle this.
Unit tests are added to cover the various cases for the helper function.