gitindex: set protocol.file.allow just in test
This makes it so you can run submodule tests without first adjusting
your global environment to allow file clones. This was a somewhat recent
change in git to require opting in. Previously we would just set the
config in CI, but now we just always do it _just_ for that test run.

Test Plan: go test and CI