list: add indextime to MinimalRepoListEntry
We want this information in Sourcegraph. This turned into a bit more of
a hairy change due to our custom marshalling (for perf) and grpc
validation.

I am strongly considering making the type for
MinimalRepoListEntry.IndexTime just an int64 and storing the unix time
without milliseconds. This not only changes the struct size from 56
bytes to 40, but likely makes much of the code simpler to reason about.

Test Plan: lots of unit test coverage