indexserver: include index time when pushing sg index updates
We missed this path in the last PR, since we assumed we did an RPC to
get this information. This change has a slightly larger blast radius
because indexserver needs to read back the index file to find out the
index time. Previously it just used inputs to the index when informing
sourcegraph.

Test Plan: pretty robust unit tests here.