RFC: introduce stats in return from ngramIndex.Get
We want to start tracking the work done by the btree. This field
introduces a stats type with the btree in mind that will be returned
anytime it is asked to do work. The intention is aggregate these values
inside of the work done by our ngram iterator construction.

This is a little different to how we do it with iterators which instead
are created per shard search and are then asked to updateStats. That
might be a nice future design, introduce a per request accessor to
indexData which can record stats. But that is a much larger change.

Note: this change has some commented out code such that noMatchTree will
also update stats. The intention is when we rule a tree out we can
capture the stats. There is still a complication around how we integrate
that with pruneMatchTree.